Add user following

+++ app/Http/Controllers/StatusesController.php
@@ -1,7 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Contracts\Auth\Access\Gate as GateContract;
 use Illuminate\Http\Request;
 use App\Models\Status;
 use App\Http\Requests;
@@ -25,8 +26,8 @@ class StatusesController extends Controller
     }
    public function destroy(Status $status)
     {
-        error_log(print_r($status,true),3,"/tmp/xinjiahui.log");
-        //$this->authorize('destroy', $status);
+       error_log(print_r($status,true),3,"/tmp/xinjiahui.log");
+        $this->authorize('destroy', $status);
         $status->delete();
         session()->flash('success', '微博已被成功删除！');
         return redirect()->back();
